requirements.txt
fastapi[standard]
sqlmodel
pyPodcastParser
requests
python-dateutil
apscheduler
alembic
freezun